requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
b2sdk==1.21.0
billiard==4.1.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-redis==5.3.0
djangorestframework==3.14.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
kombu==5.3.1
logfury==1.0.1
msgpack==1.0.5
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
